slidems/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel", "Cython"]
build-backend = "setuptools.build_meta"


[project.urls]
Homepage = "https://github.com/dudi-l/biopsyfocus"


[project]
name = "slidems"
description = "Yofi Biopsy POC"
authors = [
  {name = "Dudi Levi"},
  {email = "dudilv@gmail.com" }
]
readme = "README.md"
#license = {file = "License"}
requires-python = ">=3.11"

dynamic = ["version"]

dependencies = [
    "fastai",
    "fastcore",
    "ipywidgets",
    "jmd_imagescraper",
    "fastbook",
    "jupyterlab_widgets",
    "nbdev",
    "openslide-python",
    "matplotlib",
    "jinja2",
    "orjson"
]


[project.optional-dependencies]
development = [
    "jupyterlab",
    "pytest"
]


[tool.setuptools.packages.find]
include = ["slidems*"]
exclude = ["slidems.train"]


[tool.setuptools.package-data]
"slidems.configs.model" = ["resize512_RandomResizedCrop256_resnet50_epoc10.pkl"]



